+++ lib/airship_flutter.dart
@@ -14,7 +14,7 @@ export 'src/airship_preference_center.dart';
 export 'src/airship_privacy_manager.dart';
 export 'src/airship_push.dart';
 export 'src/airship_events.dart';
-export 'src/airship_feature_flag.dart';
+export 'src/airship_feature_flag_manager.dart';
 export 'src/push_notification_status.dart';
 export 'src/push_payload.dart';
 export 'src/ios_push_options.dart';
@@ -23,11 +23,9 @@ export 'src/attribute_editor.dart';
 export 'src/channel_scope.dart';
 export 'src/custom_event.dart';
 export 'src/feature.dart';
-export 'src/feature_flag.dart';
 export 'src/inbox_message.dart';
 export 'src/inbox_message_view.dart';
 export 'src/preference_center_config.dart';
-export 'src/push_notification_status.dart';
 export 'src/scoped_subscription_list_editor.dart';
 export 'src/subscription_list_editor.dart';
 export 'src/tag_group_editor.dart';
